refactor: Move types to own file
- Add Tailwind installer custom plugin name for debugging
